getServerMessage
GET REQUEST ASYNCHRON USING VOLLEY
