+++ Proiect Android/Interfata/app/src/main/java/com/example/myapplication/MainActivity.java
@@ -11,6 +11,13 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 
+import com.android.volley.Request;
+import com.android.volley.RequestQueue;
+import com.android.volley.Response;
+import com.android.volley.VolleyError;
+import com.android.volley.toolbox.StringRequest;
+import com.android.volley.toolbox.Volley;
+
 public class MainActivity extends AppCompatActivity {
 
 
@@ -30,4 +37,33 @@ public class MainActivity extends AppCompatActivity {
         });
 
     }
-}
+
+    public void getServerMessage(){
+        // the url is different for every computer.
+        // for emulator use 10.0.0.2:5000/
+        // for device, run ipconfig in cmd and get ipv4 address
+
+        final String url = "http://192.168.1.2:5000/";
+        final RequestQueue requestQueue = Volley.newRequestQueue(MainActivity.this);
+
+        StringRequest stringRequest = new StringRequest(Request.Method.GET, url,
+                new Response.Listener<String>() {
+                    @Override
+                    public void onResponse(String response) {
+                        //do something with response
+                    }
+                }, new Response.ErrorListener() {
+                    @Override
+                    public void onErrorResponse(VolleyError error) {
+                        //deal with errors
+                        error.printStackTrace();
+                        requestQueue.stop();
+                    }
+                });
+        requestQueue.add(stringRequest);
+
+        }
+
+    }
+
+
